requirements.txt
# ShipLock Requirements
# Production-grade dependencies for secure Docker distribution

# CLI and UI
rich>=13.0.0
click>=8.0.0

# Configuration and Data
pyyaml>=6.0

# Cryptography and Security
cryptography>=41.0.0

# Docker Integration
docker>=6.0.0

# Network and System
netifaces>=0.11.0
psutil>=5.9.0

# Development Dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
